+++ src/datasources/test-plans/constants/QueryAssets.constants.ts
@@ -1,0 +1,2 @@
+export const QUERY_ASSETS_REQUEST_PER_SECOND = 8;
+export const QUERY_ASSETS_BATCH_SIZE = 1000;